# perf: avoid extra copies when receiving int4 and int2 in PGStream