### All Submissions:

* [x] Have you followed the guidelines in our [Contributing](https://github.com/pgjdbc/pgjdbc/blob/master/CONTRIBUTING.md) document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?

### Changes to Existing Features:

This PR is for #3250 
I have moved receiveInteger2 and receiveInteger4 to VisibleBufferedInputStream. 
Within PGStream, I have also removed dependency on int4buf and int2buf by directly writing value to OutputStream inside sendInteger2 and sendInteger4
* [x] Have you successfully run tests with your changes locally?


Based on reviews, can further add tests and documentation for the new methods