@vishalvrv9 2024-06-23
@vlsi 

I have added benchmarks. After running the benchmarks on my machine, I'm getting consistent results as below. 

I'm pasting results of ```before change``` and ```after change``` for comparison

```
Before the PGStream refactoring

Benchmark                                      (byteArrSize)  Mode  Cnt         Score        Error  Units
StreamReceiveInteger.receiveInteger2Benchmark           4096  avgt   25     40654.447 ±    985.740  ns/op
StreamReceiveInteger.receiveInteger2Benchmark           8192  avgt   25     79654.366 ±   1004.068  ns/op
StreamReceiveInteger.receiveInteger2Benchmark          65536  avgt   25    642311.430 ±   6683.201  ns/op
StreamReceiveInteger.receiveInteger2Benchmark         131072  avgt   25   1276954.173 ±  18509.234  ns/op
StreamReceiveInteger.receiveInteger2Benchmark        1048576  avgt   25  10064545.268 ±  60657.317  ns/op

StreamReceiveInteger.receiveInteger4Benchmark           4096  avgt   25     21010.785 ±    157.727  ns/op
StreamReceiveInteger.receiveInteger4Benchmark           8192  avgt   25     40477.736 ±    926.200  ns/op
StreamReceiveInteger.receiveInteger4Benchmark          65536  avgt   25    316628.093 ±   3137.282  ns/op
StreamReceiveInteger.receiveInteger4Benchmark         131072  avgt   25    642833.953 ±   5306.471  ns/op
StreamReceiveInteger.receiveInteger4Benchmark        1048576  avgt   25   5259341.809 ± 123359.057  ns/op

After the refactoring


Benchmark                                      (byteArrSize)  Mode  Cnt        Score        Error  Units
StreamReceiveInteger.receiveInteger2Benchmark           4096  avgt   25    30660.816 ±    395.408  ns/op
StreamReceiveInteger.receiveInteger2Benchmark           8192  avgt   25    61470.316 ±   2574.740  ns/op
StreamReceiveInteger.receiveInteger2Benchmark          65536  avgt   25   482912.610 ±   4937.530  ns/op
StreamReceiveInteger.receiveInteger2Benchmark         131072  avgt   25   963792.516 ±   4081.982  ns/op
StreamReceiveInteger.receiveInteger2Benchmark        1048576  avgt   25  7761601.288 ± 147700.245  ns/op

StreamReceiveInteger.receiveInteger4Benchmark           4096  avgt   25    15744.789 ±     81.255  ns/op
StreamReceiveInteger.receiveInteger4Benchmark           8192  avgt   25    31209.710 ±    168.836  ns/op
StreamReceiveInteger.receiveInteger4Benchmark          65536  avgt   25   249877.018 ±   3850.310  ns/op
StreamReceiveInteger.receiveInteger4Benchmark         131072  avgt   25   501395.618 ±   5036.232  ns/op
StreamReceiveInteger.receiveInteger4Benchmark        1048576  avgt   25  4038478.701 ±  42636.446  ns/op
```
@bokken 
I have also changed the PGStream to buffer values before writing to OutputStream within sendInteger2() and sendInteger4()