@vishalvrv9 2024-06-25
@vlsi 

I have refactored the benchmarks to as to avoid reflection and directly use ```VisibleBufferedInputStream```

The results are below:

```
Benchmark                                      (byteArrSize)  Mode  Cnt        Score       Error  Units
StreamReceiveInteger.receiveInteger2Benchmark           4096  avgt   25     4902.307 ±    64.914  ns/op
StreamReceiveInteger.receiveInteger2Benchmark           8192  avgt   25     9706.844 ±    24.202  ns/op
StreamReceiveInteger.receiveInteger2Benchmark          65536  avgt   25    79798.808 ±  1115.382  ns/op
StreamReceiveInteger.receiveInteger2Benchmark         131072  avgt   25   162177.605 ±  4183.496  ns/op
StreamReceiveInteger.receiveInteger2Benchmark        1048576  avgt   25  1334331.384 ± 88054.323  ns/op

StreamReceiveInteger.receiveInteger4Benchmark           4096  avgt   25     2940.606 ±    23.006  ns/op
StreamReceiveInteger.receiveInteger4Benchmark           8192  avgt   25     5951.321 ±   111.409  ns/op
StreamReceiveInteger.receiveInteger4Benchmark          65536  avgt   25    47203.798 ±   579.416  ns/op
StreamReceiveInteger.receiveInteger4Benchmark         131072  avgt   25    95262.525 ±  1138.591  ns/op
StreamReceiveInteger.receiveInteger4Benchmark        1048576  avgt   25   774245.750 ± 14734.046  ns/op
```