@vishalvrv9 2024-06-28
Hey @vlsi 

Have implemented the proposed review changes. 

The benchmark results look as below:

```
Benchmark                                      Mode  Cnt      Score     Error  Units
StreamReceiveInteger.receiveInteger2Benchmark  avgt   25  12248.481 ± 315.238  ns/op
StreamReceiveInteger.receiveInteger4Benchmark  avgt   25   8068.027 ±  84.654  ns/op
```