@vishalvrv9 2024-08-15
Have accomodated the changes. The benchmark results are below:
```
Benchmark                                      Mode  Cnt  Score   Error  Units
StreamReceiveInteger.receiveInteger2Benchmark  avgt   25  4.829 ± 0.023  ns/op
StreamReceiveInteger.receiveInteger4Benchmark  avgt   25  4.846 ± 0.078  ns/op
```